Relocated message cfg class
