+++ worker-queue-rabbit/src/main/java/com/github/workerframework/queues/rabbit/MessageDehydrationConfiguration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.github.workerframework.configs;
+package com.github.workerframework.queues.rabbit;
 
 import jakarta.validation.constraints.Min;
