+++ worker-queue-rabbit/src/test/java/com/github/workerframework/queues/rabbit/RabbitWorkerQueuePublisherTest.java
@@ -22,7 +22,6 @@ import com.github.workerframework.api.DataStoreException;
 import com.github.workerframework.api.TaskMessage;
 import com.github.workerframework.api.TaskStatus;
 import com.github.workerframework.api.TrackingInfo;
-import com.github.workerframework.configs.MessageDehydrationConfiguration;
 import com.github.workerframework.datastores.fs.FileSystemDataStore;
 import com.github.workerframework.datastores.fs.FileSystemDataStoreConfiguration;
 import com.github.workerframework.util.rabbitmq.ConsumerRejectEvent;
@@ -41,7 +40,6 @@ import org.mockito.stubbing.Answer;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Date;
 import java.util.HashMap;
